requirements.txt
# requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pendulum==3.0.0
python-dotenv==1.0.0